LayoutTests/fast/regions/region-event-remove-from-dom-expected.txt

Tests regionLayoutEvent

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

 Flow content
PASS event.target.id is "region_1"
PASS event.target.id is "region_2"
Removing region_2 from DOM
PASS event.target.id is "region_1"
PASS successfullyParsed is true

TEST COMPLETE
